V2.0 (#188)
* 2.0 update

See changelog for more details

+++ tests/conftest.py
@@ -1,4 +1,4 @@
-from aligner.command_line.mfa import fix_path
+from montreal_forced_aligner.command_line.mfa import fix_path
 
 fix_path()
 
@@ -6,20 +6,9 @@ import os
 import shutil
 import pytest
 
-from aligner.corpus import Corpus
-from aligner.dictionary import Dictionary
-from aligner.config import train_yaml_to_config
-
-
-def pytest_addoption(parser):
-    parser.addoption("--skiplarge", action="store_true",
-                     help="skip large dataset tests")
-
-
-
-@pytest.fixture
-def skip_large(request):
-    return request.config.getoption('--skiplarge')
+from montreal_forced_aligner.corpus import AlignableCorpus
+from montreal_forced_aligner.dictionary import Dictionary
+from montreal_forced_aligner.config import train_yaml_to_config
 
 
 @pytest.fixture(scope='session')
@@ -45,6 +34,13 @@ def textgrid_dir(test_dir):
 
 @pytest.fixture(scope='session')
 def generated_dir(test_dir):
+    from montreal_forced_aligner.thirdparty.kaldi import validate_kaldi_binaries
+    if not validate_kaldi_binaries():
+        from montreal_forced_aligner.thirdparty.download import download_binaries
+        download_binaries()
+    actually_working = validate_kaldi_binaries()
+    if not actually_working:
+        raise Exception('Kaldi binaries are not correctly found or functioning.')
     generated = os.path.join(test_dir, 'generated')
     shutil.rmtree(generated, ignore_errors=True)
     if not os.path.exists(generated):
@@ -59,13 +55,21 @@ def temp_dir(generated_dir):
 
 @pytest.fixture(scope='session')
 def english_acoustic_model():
-    from aligner.command_line.download import download_model
+    from montreal_forced_aligner.command_line.download import download_model
     download_model('acoustic', 'english')
 
 
 @pytest.fixture(scope='session')
 def corpus_root_dir(generated_dir):
     return os.path.join(generated_dir, 'corpus')
+
+
+@pytest.fixture(scope='session')
+def default_feature_config():
+    from montreal_forced_aligner.features.config import FeatureConfig
+    fc = FeatureConfig()
+    fc.use_mp = False
+    return fc
 
 
 @pytest.fixture(scope='session')
@@ -103,6 +107,15 @@ def extra_corpus_dir(corpus_root_dir, wav_dir, lab_dir):
     name = 'cold_corpus3'
     shutil.copyfile(os.path.join(wav_dir, name + '.wav'), os.path.join(path, name + '.wav'))
     shutil.copyfile(os.path.join(lab_dir, name + '_extra.lab'), os.path.join(path, name + '.lab'))
+    return path
+
+
+@pytest.fixture(scope='session')
+def transcribe_corpus_24bit_dir(corpus_root_dir, wav_dir):
+    path = os.path.join(corpus_root_dir, '24bit')
+    os.makedirs(path, exist_ok=True)
+    name = 'cold_corpus_24bit'
+    shutil.copyfile(os.path.join(wav_dir, name + '.wav'), os.path.join(path, name + '.wav'))
     return path
 
 
@@ -262,7 +275,7 @@ def sick_dict(sick_dict_path, generated_dir):
 @pytest.fixture(scope='session')
 def sick_corpus(basic_corpus_dir, generated_dir):
     output_directory = os.path.join(generated_dir, 'sickcorpus')
-    corpus = Corpus(basic_corpus_dir, output_directory, num_jobs=2)
+    corpus = AlignableCorpus(basic_corpus_dir, output_directory, num_jobs=2)
     return corpus
 
 
@@ -332,6 +345,11 @@ def textgrid_output_model_path(generated_dir):
 
 
 @pytest.fixture(scope='session')
+def ivector_output_model_path(generated_dir):
+    return os.path.join(generated_dir, 'ivector_output_model.zip')
+
+
+@pytest.fixture(scope='session')
 def training_dict_path(test_dir):
     return os.path.join(test_dir, "dictionaries", "chinese_dict.txt", )
 
@@ -382,8 +400,18 @@ def basic_align_config(config_directory):
 
 
 @pytest.fixture(scope='session')
-def mono_train_config(config_directory):
-    return train_yaml_to_config(os.path.join(config_directory, 'mono_train.yaml'))
+def basic_train_ivector_config(config_directory):
+    return os.path.join(config_directory, 'basic_train_ivector.yaml')
+
+
+@pytest.fixture(scope='session')
+def mono_train_config_path(config_directory):
+    return os.path.join(config_directory, 'mono_train.yaml')
+
+
+@pytest.fixture(scope='session')
+def mono_train_config(mono_train_config_path):
+    return train_yaml_to_config(mono_train_config_path)
 
 
 @pytest.fixture(scope='session')
